requirements.txt
PyYAML>=6.0
pandas>=2.0.0
pyarrow>=12.0.0
tqdm>=4.66.0
numpy>=1.24.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.28.0
sentencepiece>=0.1.99
rouge-score>=0.1.2
